requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
dataclasses-json==0.6.7
exceptiongroup==1.3.0
fastapi==0.104.1
frozenlist==1.7.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-tools==1.73.1
h11==0.16.0
httplib2==0.22.0
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.0
langchain-community==0.0.38
langchain-core==0.1.53
langchain-google-genai==0.0.6
langgraph==0.0.26
langsmith==0.0.92
marshmallow==3.26.1
multidict==6.6.3
mypy-extensions==1.1.0
numpy==1.26.4
oauthlib==3.3.1
packaging==23.2
propcache==0.3.2
proto-plus==1.26.1
protobuf>=3.19.5,<5.0.0dev
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.5.0
pydantic-core==2.14.1
pyparsing==3.2.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.27.0
tenacity==8.5.0
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspect==0.9.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
yarl==1.20.1
